feat(shares): add SharesTourStep component and implement share management features
- Introduced SharesTourStep component for guided tour of share management functionalities.
- Created Shares page with comprehensive share management capabilities including create, edit, enable/disable, and delete shares.
- Implemented state management for share selection and dialog handling.
- Added utility functions for share name sanitization and casing transformations.
- Enhanced user experience with tooltips and icons for various share actions.
- Integrated event handling for guided tour steps to improve user onboarding.
